EPMDJ-2717: optimized package tree
Issue: EPMDJ-2717
