+++ src/components/table/index.ts
@@ -1,4 +1,3 @@
 export { Table } from './table';
 export { Column } from './column';
-export { SelectableTable } from './selectable-table';
 export { ExpandableTable } from './expandable-table';
