+++ src/components/table/table.module.scss
@@ -33,8 +33,19 @@
       background-color: rgba(104, 116, 129, 0.05);
     }
 
+    &--nested {
+      border-left: 2px solid map-get($color-palette, blue15);
+      border-right: 2px solid map-get($color-palette, blue15);
+    }
+
+    &--nested-last {
+      border-bottom: 2px solid map-get($color-palette, blue15);
+    }
+
     &--color {
       &--blue {
+        border-left: 2px solid map-get($color-palette, blue15);
+        border-right: 2px solid map-get($color-palette, blue15);
         background-color: rgba(0, 127, 255, 0.15);
 
         &:hover {
