+++ src/pages/agent-full-page/coverage-plugin/coverage-details/coverage-details.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { BEM } from '@redneckz/react-bem-helper';
 
-import { ExpandableTable, Column, Icons } from 'components';
+import { Icons } from 'components';
 import { Panel } from 'layouts';
 import { ClassCoverage } from 'types/class-coverage';
 import { CompoundCell } from './compound-cell';
@@ -9,6 +9,7 @@ import { CoverageCell } from './coverage-cell';
 import { NameCell } from './name-cell';
 import { AssociatedTestModal } from './associated-test-modal';
 import { AssociatedTestColumn } from './associated-test-column';
+import { ExpandableTable, Column } from './table';
 
 import styles from './coverage-details.module.scss';
 
@@ -17,13 +18,14 @@ interface Props {
   coverageByPackages: ClassCoverage[];
   header?: React.ReactNode;
   associatedTestsTopic: string;
+  classesTopicPrefix: string;
 }
 
 const coverageDetails = BEM(styles);
 
 export const CoverageDetails = coverageDetails(
   ({
-    className, coverageByPackages, header, associatedTestsTopic,
+    className, coverageByPackages, header, associatedTestsTopic, classesTopicPrefix,
   }: Props) => {
     const [selectedId, setSelectedId] = React.useState('');
     return (
@@ -35,6 +37,7 @@ export const CoverageDetails = coverageDetails(
               data={coverageByPackages}
               idKey="name"
               columnsSize="medium"
+              classesTopicPrefix={classesTopicPrefix}
               expandedColumns={[
                 <Column
                   name="name"
@@ -71,7 +74,7 @@ export const CoverageDetails = coverageDetails(
                   Cell={(props) => <AssociatedTestColumn onClick={setSelectedId} {...props} />}
                 />,
               ]}
-              expandedContentKey="classes"
+              expandedContentKey="name"
               hasSecondLevelExpand
             >
               <Column
