+++ src/pages/agent-full-page/coverage-plugin/coverage-details/table/default-header-cell.tsx
