+++ src/pages/agent-full-page/coverage-plugin/coverage-details/table/expandable-table/index.ts
