+++ src/pages/agent-full-page/coverage-plugin/coverage-details/table/expandable-table/row-expander/row-expander.tsx
