+++ src/pages/agent-full-page/coverage-plugin/coverage-details/table/index.ts
