+++ src/pages/agent-full-page/coverage-plugin/coverage-details/table/table.tsx
@@ -17,9 +17,11 @@ interface Props {
   columnsSize?: 'wide' | 'medium';
   expandedRows?: string[];
   expandedColumns?: any[];
+  secondLevelExpand?: any[];
   expandedContentKey?: string;
   withoutHeader?: boolean;
   selectedRows?: string[];
+  classesTopicPrefix: string;
 }
 
 const table = BEM(styles);
@@ -34,12 +36,18 @@ export const Table = table(
     expandedRows = [],
     expandedColumns,
     expandedContentKey,
+    secondLevelExpand,
     withoutHeader,
     selectedRows = [],
+    classesTopicPrefix,
   }: Props) => {
     const columns = React.Children.map(children, (column) => column && column.props);
     const expandedColumnsComponents = React.Children.map(
       expandedColumns,
+      (column) => column && column.props,
+    );
+    const expandedColumnsSecondLevel = React.Children.map(
+      secondLevelExpand,
       (column) => column && column.props,
     );
 
@@ -57,6 +65,8 @@ export const Table = table(
               color={getRowColor({ expandedRows, selectedRows, itemId: String(item[idKey]) })}
               expandedContentKey={expandedContentKey}
               expandedRows={expandedRows}
+              secondLevelExpand={expandedColumnsSecondLevel}
+              classesTopicPrefix={classesTopicPrefix}
             />
           ))}
         </tbody>
