+++ src/pages/agent-full-page/coverage-plugin/coverage-details/table/table__header.tsx
